+++ src/Ens.ts
@@ -1,7 +1,7 @@
 import { default as ensInterface } from './contracts/ens/ens';
 import { default as resolverInterface } from './contracts/ens/resolver';
 import { formatsByCoinType } from '@ensdomains/address-encoder';
-import { EthCoinIndex, Bip44Constants, BlockhanNetworkUrlMap, NetworkIdMap } from './types';
+import { EthCoinIndex, Bip44Constants, BlockhanNetworkUrlMap, NetworkIdMap } from './types/types';
 import {
   NamingServiceName,
   ResolutionError,
@@ -10,7 +10,7 @@ import {
 import Contract from './utils/contract';
 import contentHash from 'content-hash';
 import EnsNetworkMap from 'ethereum-ens-network-map';
-import { Provider, EnsSource, EnsConfig } from './publicTypes';
+import { Provider, EnsSource, EnsConfig } from './types/publicTypes';
 import { buildContract, constructRecords, ensureConfigured, invert, isNullAddress } from './utils';
 import NamingService from './interfaces/NamingService';
 import FetchProvider from './FetchProvider';
