+++ src/FetchProvider.ts
@@ -1,7 +1,7 @@
 import { FetchError } from 'node-fetch';
-import { RequestArguments } from './types';
+import { RequestArguments } from './types/types';
 import ResolutionError, { ResolutionErrorCode } from './errors/resolutionError';
-import { ResolutionMethod, Provider } from './publicTypes';
+import { ResolutionMethod, Provider } from './types/publicTypes';
 import Networking from './utils/Networking';
 
 export default class FetchProvider implements Provider {
