+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Resolution, Resolution as default } from './Resolution';
-export * from './publicTypes';
+export * from './types/publicTypes';
 export { default as DnsUtils } from './utils/DnsUtils';
 export { ResolutionError, ResolutionErrorCode } from './errors/resolutionError';
 export {
