+++ src/utils/Eip1993Factories.ts
@@ -1,4 +1,4 @@
-import { RequestArguments } from '../types';
+import { RequestArguments } from '../types/types';
 import {
   ConfigurationError,
   ConfigurationErrorCode,
@@ -12,7 +12,7 @@ import {
   Web3Version1Provider,
   JsonRpcResponse,
   EthersProvider,
-} from '../publicTypes';
+} from '../types/publicTypes';
 
 export const Eip1993Factories = {
   fromWeb3Version0Provider,
