+++ src/utils/contract.ts
@@ -1,6 +1,6 @@
 import { Interface, JsonFragment } from '@ethersproject/abi';
-import { RequestArguments, EventData } from '../types';
-import { Provider } from '../publicTypes';
+import { RequestArguments, EventData } from '../types/types';
+import { Provider } from '../types/publicTypes';
 
 export default class Contract {
   readonly abi: JsonFragment[];
