+++ src/utils/index.ts
@@ -1,7 +1,7 @@
 import ConfigurationError, { ConfigurationErrorCode } from '../errors/configurationError';
 import ResolutionError, { ResolutionErrorCode } from '../errors/resolutionError';
-import { CryptoRecords, TickerVersion, EnsSupportedNetworks, CnsSupportedNetworks, Provider, ZnsSupportedNetworks, NamingServiceName, Api } from '../publicTypes';
-import { isCnsSupportedNetworks, NullAddresses, isEnsSupportedNetworks, isZnsSupportedNetworks } from '../types';
+import { CryptoRecords, TickerVersion, EnsSupportedNetworks, CnsSupportedNetworks, Provider, ZnsSupportedNetworks, NamingServiceName, Api } from '../types/publicTypes';
+import { isCnsSupportedNetworks, NullAddresses, isEnsSupportedNetworks, isZnsSupportedNetworks } from '../types/types';
 import Contract from './contract';
 
 /**
